Some refactorings
