+++ plugins/de.darwinspl.feature.graphical.configurator/META-INF/MANIFEST.MF
@@ -47,7 +47,8 @@ Require-Bundle: org.eclipse.ui,
  org.deltaecore.suite.mapping,
  org.eclipse.emf.common,
  de.darwinspl.solver,
- eu.hyvar.context.information
+ eu.hyvar.context.information,
+ de.darwinspl.variant_generation
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: www.darwinspl.de
