+++ plugins/de.darwinspl.variant_generation/src/de/darwinspl/variant_generation/impl/VariantDerivationImpl.java
@@ -27,6 +27,10 @@ public class VariantDerivationImpl implements VariantDerivation {
 	}
 
 	@Override
+	@Deprecated
+	/**
+	 * Needs to be updated to new varieant derivation!
+	 */
 	public void deriveVariant(HyFeatureModel featureModel, HyConstraintModel constraints, HyConfiguration configuration,
 			HyMapping mapping, Date date) throws VariantDerivationFailedException {
